+++ migrations/env.py
@@ -1,8 +1,11 @@
 from __future__ import annotations
+
 import os
 from logging.config import fileConfig
+
 from alembic import context
 from sqlalchemy import engine_from_config, pool
+
 from src.infrastructure.database.models import Base
 
 config = context.config